flask-consul/requirements.txt
Werkzeug==2.0.3
itsdangerous==2.0.1
flask==2.0.2
flask-restful==0.3.9
flask-cors==3.0.10
Flask-HTTPAuth==4.5.0
requests==2.27.1
Flask-APScheduler==1.12.3
#pyDes==2.0.1
pycryptodome==3.14.1
huaweicloudsdkcore==3.0.78
huaweicloudsdkecs==3.0.78
huaweicloudsdkeps==3.0.78
alibabacloud_resourcemanager20200331==2.1.0
alibabacloud_ecs20140526==2.1.0
tencentcloud-sdk-python-common==3.0.607
tencentcloud-sdk-python-cvm==3.0.607
tencentcloud-sdk-python-dcdb==3.0.607
